Add -summary, to spit out a single line result for each election to stdout
Also, fixed handling of write-in totals and percentages when none exist.
It appears in such a case, the clerk's software simply doesn't print anything
for write-ins in the total or percentage field, so you can't use positional
calculations from the back of the list to slice off the list. Instead,
start at the beginning of the list, look for the first number, and pad it
out with zeros to the end